# WARNING: head commit changed in the meantime
expectimax finished

+++ src/ExpectiMax/MaxNode.java
@@ -4,14 +4,16 @@ import main.Move2;
 
 public class MaxNode extends ExpectiNode
 {
-
-	public MaxNode(int score, int playerID)
+	
+	public MaxNode(double score, int playerID)
 	{
 		super(score, playerID);
 	}
 	
-	public MaxNode(ExpectiNode parent, int score, int playerID, Move2 move) {
+	public MaxNode(ExpectiNode parent, double score, int playerID, Move2 move) {
 		super(parent, score, playerID, move);
 	}
+	
+	
 	
 }
