+++ src/ExpectiMax/MinNode.java
@@ -5,12 +5,12 @@ import main.Move2;
 public class MinNode extends ExpectiNode
 {
 
-	public MinNode(int score, int playerID)
+	public MinNode(double score, int playerID)
 	{
 		super(score, playerID);
 	}
 	
-	public MinNode(ExpectiNode parent, int score, int playerID, Move2 move) 
+	public MinNode(ExpectiNode parent, double score, int playerID, Move2 move) 
 	{
 		super(parent, score, playerID, move);
 	}
